+++ routes/users.router.js
@@ -7,6 +7,9 @@ const router = express.Router()
 
 router.get('/', async (req, res, next) => {
 	try {
+		if (!req.token.permissions.users?.includes('list')) {
+			throw new UnauthorizedError("Você não tem permissão para acessar essa página")
+		}
 		result = await User.find({}, { password: 0 })
 		res.json(result)
 	} catch (error) {
@@ -17,6 +20,9 @@ router.get('/', async (req, res, next) => {
 
 router.get('/:id', async (req, res, next) => {
 	try {
+		if (!req.token.permissions.users?.includes('list')) {
+			throw new UnauthorizedError("Você não tem permissão para acessar essa página")
+		}
 		const id = req.params.id
 		const result = await User.findById(id, { password: 0 }).populate('group').populate('list')
 		if (!result)
@@ -32,6 +38,9 @@ router.get('/:id', async (req, res, next) => {
 
 router.delete('/:id', async (req, res, next) => {
 	try {
+		if (!req.token.permissions.users?.includes('delete')) {
+			throw new UnauthorizedError("Você não tem permissão para acessar essa página")
+		}
 		const id = req.params.id
 		const result = await User.findOneAndDelete({ _id: id })
 		if (!result)
@@ -47,6 +56,9 @@ router.delete('/:id', async (req, res, next) => {
 
 router.patch('/:id', async (req, res, next) => {
 	try {
+		if (!req.token.permissions.users?.includes('edit')) {
+			throw new UnauthorizedError("Você não tem permissão para acessar essa página")
+		}
 		const id = req.params.id
 		req.body.updatedAt = Date.now()
 		const result = await User.findOneAndUpdate({ _id: id }, req.body)
